Work with XSLT files on disk rather than from JAR
Two extra parameters give the location of the 'common' source
XSLT files for the tree transforms, and of the 'specific' XSLT
encoder generated for the user schema.

Also, a temporary folder is now used with a temporary file
created for each intermediate step, instead of a single temporary
file for the final output previously.

These are two steps away from the initial intent for the
standalone encoder, which, at the same time, make the code
all the more complicated. This commit should be reverted.